requirements.txt
blinker==1.9.0
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
Flask==3.1.2
fonttools==4.61.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
nltk==3.9.2
numpy==2.4.2
packaging==26.0
pandas==3.0.0
pillow==12.1.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
regex==2026.1.15
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
six==1.17.0
threadpoolctl==3.6.0
tqdm==4.67.3
Werkzeug==3.1.5
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.0.0
torch>=2.0.0
transformers>=4.0.0
accelerate>=1.1.0
python-dotenv>=1.0.0
openenv>=0.1.0
openai>=1.0.0
pytest>=8.0.0
httpx>=0.27.0
